Adapatar gerador de relatoriaos usando consulta mysql

+++ collecta/src/main/java/collectiva/org/collecta/domain/relatorio/Relatorio.java
@@ -22,9 +22,9 @@ public class Relatorio {
     private BigDecimal valorArrecadado;
     private BigDecimal valorRestante;
     private int visualizacoes;
-    private int contribuicoesMonetarias;
-    private int contribuicoesRecurso;
-    private int contribuicoesServicos;
+    private Long contribuicoesMonetarias;
+    private Long contribuicoesRecurso;
+    private Long contribuicoesServicos;
     private LocalDateTime data;
 
     @ManyToOne
