+++ collecta/src/main/java/collectiva/org/collecta/domain/relatorio/dto/ResponseRelatorioDTO.java
@@ -15,8 +15,8 @@ public class ResponseRelatorioDTO {
     private BigDecimal valorArrecadado;
     private BigDecimal valorRestante;
     private int visualizacoes;
-    private int contribuicoesMonetarias;
-    private int contribuicoesRecurso;
-    private int contribuicoesServicos;
+    private Long contribuicoesMonetarias;
+    private Long contribuicoesRecurso;
+    private Long contribuicoesServicos;
     private LocalDateTime data;
 }
